paypal: pass discount information on a per-item basis
In a way, this reverts commit 1237:9838081afd30 ("Pass item discounts to
PayPal, by unrolling list of discounted items"), where we squashed the
item discount into the item price passed to paypal.

By passing discount information on a per-item basis rather than for the
whole cart (order.item_discounts), the paypal bill generated displays
the discounts in a pretty way on a per-item basis.

After this, there are no more users of the 'flatten_discount' filter,
but keep the filter and satchmo_itemlist.py filter; people might be
using it.