paypal.views.confirm_info: check if subscription product module is installed
Although right now this code block is safe even if the subscription
product module isn't installed, this change saves us the trouble of
looping through the order's items if the module isn't installed,
yielding a slight performance advantage.

order_items is now defined only when this code is run - this is safe, as
only this code block uses it.

Also, reverse the product.is_subscription check to reduce the level of
indentation necessary.