Bugfix: Replaced "self.__class__" in form signals by explicit class
This should avoid problems of double-sended signals when form classes are
derived from this classes. The first signal would be sent for the parent
class with the sender set to the child class. This may lead to problems
in signal listeners because the child class is not yet fully initialised
at this stage.